Update unit_quality_gui.py
I changed save_path to somewhere else outside the kilosort folder, and bombcell GUI is not able to find it. I modified the code so that now bc.unit_quality_gui takes an additional argument "save_path". If save_path is None, it will use ks_path as the save_path. Then, this is feed into InteractiveUnitQualityGUI.